+++ ingenialink/register.py
@@ -3,6 +3,7 @@ from typing import Any, Dict, List, Optional, Tuple, Union
 
 from ingenialink import exceptions as exc
 from ingenialink.enums.register import REG_ACCESS, REG_ADDRESS_TYPE, REG_DTYPE, REG_PHY
+from ingenialink.utils._utils import convert_bytes_to_dtype
 
 dtypes_ranges: Dict[REG_DTYPE, Dict[str, Union[int, float]]] = {
     REG_DTYPE.U8: {"max": 255, "min": 0},
@@ -36,6 +37,8 @@ class Register(ABC):
         scat_id: Sub-category ID.
         internal_use: Internal use.
         address_type: Address tpye.
+        description: Register description.
+        default: Register default value.
 
     Raises:
         TypeError: If any of the parameters has invalid type.
@@ -63,6 +66,8 @@ class Register(ABC):
         scat_id: Optional[str] = None,
         internal_use: int = 0,
         address_type: Optional[REG_ADDRESS_TYPE] = None,
+        description: Optional[str] = None,
+        default: Optional[bytes] = None,
     ) -> None:
         if labels is None:
             labels = {}
@@ -86,6 +91,8 @@ class Register(ABC):
         self._internal_use = internal_use
         self._storage_valid = False if not storage else True
         self._address_type = address_type
+        self._description = description
+        self._default = default
         self._enums = enums
         self.__config_range(reg_range)
 
@@ -236,6 +243,18 @@ class Register(ABC):
     def address_type(self) -> Optional[REG_ADDRESS_TYPE]:
         """Address type of the register."""
         return REG_ADDRESS_TYPE(self._address_type)
+
+    @property
+    def description(self) -> Optional[str]:
+        """Register description."""
+        return self._description
+
+    @property
+    def default(self) -> Union[None, int, float, str]:
+        """Register default value"""
+        if self._default is None:
+            return self._default
+        return convert_bytes_to_dtype(self._default, self.dtype)
 
     @property
     def mapped_address(self) -> int: